test: move testing utility functions into their own package
