+++ test/testutil/constants.go
@@ -1,0 +1,7 @@
+package testutil
+
+const (
+	DefaultCrdName = "samplecrd"
+	DefaultCrName  = "samplecr"
+	DefaultGroup   = "e2e.io"
+)
